+++ NEWS.md
@@ -1,3 +1,10 @@
+# WDI 2.7.2
+
+* Add support for several languages with the "language" argument in `WDI()` (#15)
+* Add "latest" argument in `WDI()`, to fetch the most recent non-NA values (#29, #43)
+* Speed improvement in WDIbulk (Thanks to @etiennebacher)
+* WDIbulk gets a `timeout` argument
+
 # WDI 2.7.1
 
 * Bug: did not download all countries